+++ app/components/nav_bar/user_component.html.erb
@@ -24,13 +24,13 @@
                   <% end %>
                   <%# For businesses with developer profiles %>
                   <% if business? %>
-                    <%= link_to new_businesses_form_path, class: "hidden md:block text-gray-300 hover:bg-gray-700 hover:text-white px-3 py-2 rounded-md text-sm font-medium" do %>
+                    <%= link_to new_forms_businesses_hire_path, class: "hidden md:block text-gray-300 hover:bg-gray-700 hover:text-white px-3 py-2 rounded-md text-sm font-medium" do %>
                       <span class="absolute -mt-2.5 -ml-6 z-10 uppercase text-indigo-400 tracking-tight font-bold text-xs -rotate-12"><%= t(".new") %></span>
                       <span><%= t(".hired_a_dev") %></span>
                     <% end %>
                   <% end %>
                 <% elsif business? %>
-                  <%= link_to new_businesses_form_path, class: "hidden md:block text-gray-300 hover:bg-gray-700 hover:text-white px-3 py-2 rounded-md text-sm font-medium" do %>
+                  <%= link_to new_forms_businesses_hire_path, class: "hidden md:block text-gray-300 hover:bg-gray-700 hover:text-white px-3 py-2 rounded-md text-sm font-medium" do %>
                     <span class="absolute -mt-2.5 -ml-6 z-10 uppercase text-indigo-400 tracking-tight font-bold text-xs -rotate-12"><%= t(".new") %></span>
                     <span><%= t(".hired_a_dev") %></span>
                   <% end %>
@@ -124,13 +124,13 @@
             <span class="absolute -mt-2.5 z-10 uppercase text-indigo-400 tracking-tight font-bold text-xs rotate-12"><%= t(".new") %></span>
           <% end %>
           <% if business? %>
-            <%= link_to new_businesses_form_path, class: "text-gray-300 hover:bg-gray-700 hover:text-white block px-3 py-2 rounded-md text-base font-medium" do %>
+            <%= link_to new_forms_businesses_hire_path, class: "text-gray-300 hover:bg-gray-700 hover:text-white block px-3 py-2 rounded-md text-base font-medium" do %>
               <span><%= t(".hired_a_dev") %></span>
               <span class="absolute -mt-2.5 z-10 uppercase text-indigo-400 tracking-tight font-bold text-xs rotate-12"><%= t(".new") %></span>
             <% end %>
           <% end %>
         <% elsif business? %>
-          <%= link_to new_businesses_form_path, class: "text-gray-300 hover:bg-gray-700 hover:text-white block px-3 py-2 rounded-md text-base font-medium" do %>
+          <%= link_to new_forms_businesses_hire_path, class: "text-gray-300 hover:bg-gray-700 hover:text-white block px-3 py-2 rounded-md text-base font-medium" do %>
             <span><%= t(".hired_a_dev") %></span>
             <span class="absolute -mt-2.5 z-10 uppercase text-indigo-400 tracking-tight font-bold text-xs rotate-12"><%= t(".new") %></span>
           <% end %>
